requirements.txt
langchain
langchain-core
langchain-openai
langchain-community
langgraph
supabase
python-dotenv
